in index.js add component App. Add useSelector for receiving data
